refactor: 기능 단위로 메소드 분리
